// ui/src/app/@api-module/model/purchase.ts
/**
 * Purchases API
 * Optional multiline or single-line description in [CommonMark](http://commonmark.org/help/) or HTML.
 *
 * The version of the OpenAPI document: 0.1.9
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PurchaseItem } from './purchaseItem';


export interface Purchase { 
    id?: number;
    createdDate?: Date;
    checked?: boolean;
    checkedDate?: Date;
    name?: string;
    text?: string;
    username?: string;
    shared?: boolean;
    sharedForUsername?: string;
    items?: Array<PurchaseItem>;
}